Added a noframe template for showing surveys on ballons on the map

+++ coralwatch-webapp/src/main/java/org/coralwatch/dataaccess/SurveyRatingDao.java
@@ -13,5 +13,5 @@ public interface SurveyRatingDao extends Dao<SurveyRating> {
 
     double getCommunityRatingValue(Survey survey);
 
-    int getNumberOfRatingsForASurvey(Survey survey);
+    long getNumberOfRatingsForASurvey(Survey survey);
 }
